refactor(tree_attn): decouple FSDP and Megatron implementations
Split module.py into separate files to allow FSDP engine to import
tree attention functionality without requiring Megatron dependencies:

- constants.py: shared BLOCK_SIZE constant (lightweight, no deps)
- module_fsdp.py: FSDP-only functionality (PyTorch only)
- module_megatron.py: Megatron-specific functionality (mbridge, megatron.core)
- module.py: backward-compatible re-exports with conditional Megatron imports
