+++ areal/models/tree_attn/constants.py
@@ -1,0 +1,5 @@
+"""Shared constants for tree attention functionality."""
+
+import os
+
+BLOCK_SIZE = int(os.environ.get("AREAL_FLEX_ATTENTION_BLOCK_SIZE", "128"))
